% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_tools.R
\name{analyse_modele}
\alias{analyse_modele}
\title{glm model summary}
\usage{
analyse_modele(modele, train_response)
}
\arguments{
\item{modele}{Object type glm.}

\item{train_response}{La variable reponse du train (spam de type facteur).}
}
\value{
List of 3 elements : boolean, num et data frame.
}
\description{
Quelques infos sur le modele.
La convergence du modele, le R2 de Mc Fadden et la distribution $p_{xi}$ \code{...}
}
