% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_tools.R
\name{conf_mat}
\alias{conf_mat}
\title{Matrice de confusion.}
\usage{
conf_mat(var_response, var_estime, nb_individus)
}
\arguments{
\item{var_response}{Vecteur de la variable reponse sur le test (type factor pour \code{\link[=table]{table()}}).}

\item{var_estime}{Vecteur des Y estimés sur test.}

\item{nb_individus}{Longueur du data frame test (nombre reel).}
}
\value{
Data frame.
}
\description{
Sensibilite, specificite \code{...}
}
\seealso{
Other Analyse modele glm_logit.: 
\code{\link{analyse_modele}()},
\code{\link{ft_conf_tab}()}
}
\concept{Analyse modele glm_logit.}
