% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_tools.R
\name{conf_mat_rpart}
\alias{conf_mat_rpart}
\title{Matrice de confusion pour rpart.}
\usage{
conf_mat_rpart(var_response, class_estime, nb_individus)
}
\arguments{
\item{var_response}{Vecteur de la variable reponse sur le test (type factor pour \code{\link[=table]{table()}}).}

\item{class_estime}{Vecteur (factor) des Y estimés sur test.}

\item{nb_individus}{Longueur du data frame test (nombre reel).}
}
\value{
Data frame.
}
\description{
Sensibilite, specificite \code{...}
}
