% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_tools.R
\name{cv_glm_net}
\alias{cv_glm_net}
\title{Cross-validation for glmnet}
\usage{
cv_glm_net(data_df, y_name, x_name, alpha_, ...)
}
\arguments{
\item{data_df}{Model data, class "data.frame".}

\item{y_name}{The variable cible (Y) in characters.}

\item{x_name}{List of explanatory variables (list of characters).}

\item{alpha_}{0 for Ridge and 1 for Lasso.}

\item{...}{Add arguments from \code{\link[=cv.glmnet]{cv.glmnet()}}.}
}
\description{
Does k-fold cross-validation for glmnet, produces a plot \code{...}
By default, nfolds = 10
}
\seealso{
\code{\link[=cv.glmnet]{cv.glmnet()}} for initial function

Other modeling functions: 
\code{\link{glm_logit}()},
\code{\link{glm_net}()},
\code{\link{tree_rpart}()}
}
\concept{modeling functions}
