% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_tools.R
\name{def_model_keras}
\alias{def_model_keras}
\title{Define neuronal model with Keras}
\usage{
def_model_keras(
  input_neuro,
  nb__neuro_cc,
  fct_activ,
  name_loss,
  name_optimizer,
  name_metrics
)
}
\arguments{
\item{input_neuro}{Number to define "input shape".}

\item{nb__neuro_cc}{Number of neurons of blind layer.}

\item{fct_activ}{Name of activation function.}

\item{name_loss}{Name of loss function.}

\item{name_optimizer}{Name of optimizer parameter.}

\item{name_metrics}{Name of metrics.}
}
\description{
Type of model is sequential.
}
\seealso{
\code{\link[=keras_model_sequential]{keras_model_sequential()}}, \code{\link[=compile.keras.engine.training.Model]{compile.keras.engine.training.Model()}}.

Other modeling functions: 
\code{\link{cv_glm_net}()},
\code{\link{fit_keras}()},
\code{\link{glm_logit}()},
\code{\link{glm_net}()},
\code{\link{tree_rpart}()}
}
\concept{modeling functions}
