% HEADER %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% This is file spam/man/defunct.Rd.                         %
% It is part of the R package spam,                         %
%  --> https://CRAN.R-project.org/package=spam              %
%  --> https://CRAN.R-project.org/package=spam64            %
%  --> https://git.math.uzh.ch/reinhard.furrer/spam         %
% by Reinhard Furrer [aut, cre], Florian Gerber [aut],      %
%    Roman Flury [aut], Daniel Gerber [ctb],                %
%    Kaspar Moesinger [ctb]                                 %
% HEADER END %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{spam-defunct}
\alias{spam-defunct}
\alias{validspamobject}
\title{Defunct Objects in Package \pkg{spam}}
\description{
  The functions or variables listed here are defunct, i.e. thorw an error when used.
}
\usage{
  validspamobject(...)
}
\arguments{
  \item{...}{some arguments}
}
\seealso{
  \code{\link{Deprecated}},
  \code{\link{Defunct}}
}
\keyword{misc}
