% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_tools.R
\name{ft_conf_tab}
\alias{ft_conf_tab}
\title{Donnees de la matrice de confusion \code{\link[=flextable]{flextable()}}.}
\usage{
ft_conf_tab(df_conf_mat)
}
\arguments{
\item{df_conf_mat}{Data frame.}
}
\value{
Un tableau en image (viewer) et un objet de type list().
}
\description{
Pour affichage des valeurs dans un tableau.
}
\seealso{
Other Analyse modele glm_logit.: 
\code{\link{analyse_modele}()},
\code{\link{conf_mat}()}
}
\concept{Analyse modele glm_logit.}
