% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{gg_bar}
\alias{gg_bar}
\title{Proportion Email/Spam.}
\usage{
gg_bar(df, y_plot, labs_capt, labs_title)
}
\arguments{
\item{df}{Data frame object.}

\item{y_plot}{Object, name of column of your data frame.}

\item{labs_capt}{Character label.}

\item{labs_title}{Character label.}
}
\value{
ggplot/geom_bar/theme_minimal/theme/labs.
}
\description{
Barplot volume + poucentage.

Data input is data frame.
}
\seealso{
Other graph stat des: 
\code{\link{gg_bar_group}()},
\code{\link{gg_tile}()}
}
\concept{graph stat des}
