% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{gg_bar_group}
\alias{gg_bar_group}
\title{Average bar plot by group Email/Spam.}
\usage{
gg_bar_group(tidy_df, x_plot, y_plot, fill_plot, labs_capt, labs_title)
}
\arguments{
\item{tidy_df}{Data frame object.}

\item{x_plot}{Object, name of column of your data frame.}

\item{y_plot}{Object, name of column of your data frame (response variable).}

\item{fill_plot}{Object, name of column of your data frame (value).}

\item{labs_capt}{Character label.}

\item{labs_title}{Character label.}
}
\value{
ggplot/geom_bar/theme_minimal/theme/labs.
}
\description{
Barplot, numbers of capital letters (average, longest, total).

Data (tidy format) input is data frame.
}
