% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{gg_keras}
\alias{gg_keras}
\title{Plot keras training model}
\usage{
gg_keras(keras_histo, labs_capt, labs_title)
}
\arguments{
\item{keras_histo}{keras_training_history object class.}

\item{labs_capt}{Character label.}

\item{labs_title}{Character label.}
}
\value{
ggplot/theme/labs.
}
\description{
You can see, evolution of loss and accuracy.
}
