% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{gg_pi}
\alias{gg_pi}
\title{Distribution des  \code{$p_{xi}$}}
\usage{
gg_pi(df, pi, fill_var, labs_capt, labs_title)
}
\arguments{
\item{df}{Put data frame.}

\item{pi}{Name of object corresponding at estimate values of your data frame.}

\item{fill_var}{Name of object corresponding at your response variable of your data frame.}

\item{labs_capt}{Character label.}

\item{labs_title}{Character label.}
}
\value{
ggplot/geom_histogram/theme_minimal/theme/labs.
}
\description{
A utiliser apres la fonction \code{\link[=analyse_modele]{analyse_modele()}}.
}
