% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{gg_roc}
\alias{gg_roc}
\title{Courbe ROC GLM}
\usage{
gg_roc(roc, labs_capt, labs_title)
}
\arguments{
\item{roc}{Put roc object \code{\link[=roc]{roc()}}.}

\item{labs_capt}{Character label.}

\item{labs_title}{Character label.}
}
\value{
ggplot/ggroc/theme_minimal/theme/labs.
}
\description{
Permet de visualiser la qualité du pouvoir predictif du modele.
}
\seealso{
Other graphiques modele glm_logit: 
\code{\link{gg_pi}()}
}
\concept{graphiques modele glm_logit}
