% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{gg_rpart}
\alias{gg_rpart}
\title{Dendrogram with ggplot2}
\usage{
gg_rpart(modele, labs_title, labs_capt)
}
\arguments{
\item{modele}{Rpart object type.}

\item{labs_title}{Character label.}

\item{labs_capt}{Character label.}
}
\value{
ggplot/geom_segment/geom_text/theme_dendro/theme/labs.
}
\description{
Plot rpart tree model with {ggplot2} and {ggdendro}.
}
