% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{gg_tile}
\alias{gg_tile}
\title{Analyse descriptive avec geom_tile().}
\usage{
gg_tile(
  tidy_df,
  x_plot,
  y_plot,
  fill_plot,
  labs_x,
  labs_y,
  labs_capt,
  labs_title
)
}
\arguments{
\item{tidy_df}{Tidy data frame (id, nom_col, values).}

\item{x_plot}{Object, name of column of your tidy df to plot on x axis.}

\item{y_plot}{Object, name of column of your tidy df to plot on y axis.}

\item{fill_plot}{Object, name of column of your tidy df to scale.}

\item{labs_x}{Character label.}

\item{labs_y}{Character label.}

\item{labs_capt}{Character label.}

\item{labs_title}{Character label.}
}
\value{
ggplot/geom_tile/theme_minimal/them/labs.
}
\description{
Visualiser rapidement des concentrations (scale).

Data input is "tidy" data frame.
}
\seealso{
Other graph stat des: 
\code{\link{gg_bar_group}()},
\code{\link{gg_bar}()}
}
\concept{graph stat des}
