% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_tools.R
\name{glm_logit}
\alias{glm_logit}
\title{glm model}
\usage{
glm_logit(data_df, y_name, x_name, ...)
}
\arguments{
\item{data_df}{Model data, class "data.frame".}

\item{y_name}{The variable cible (Y) in characters.}

\item{x_name}{List of explanatory variables (list of characters).}

\item{...}{Add arguments from \code{\link[=glm]{glm()}}.}
}
\description{
Fitting Generalized Linear Models.
Glm is used to fit generalized linear models \code{...}
}
\examples{
nom_col <- setdiff(names(iris), "Species")
glm_logit(iris, "Species", nom_col)
}
\seealso{
\code{\link[=glm]{glm()}} for initial function

Other modeling functions: 
\code{\link{cv_glm_net}()},
\code{\link{def_model_keras}()},
\code{\link{fit_keras}()},
\code{\link{glm_net}()},
\code{\link{tree_rpart}()}
}
\concept{modeling functions}
