% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_tools.R
\name{glm_net}
\alias{glm_net}
\title{fit a GLM with Lasso or Ridge or elasticnet regularization}
\usage{
glm_net(data_df, y_name, x_name, alpha_, lambda_cv, ...)
}
\arguments{
\item{data_df}{Model data, class "data.frame".}

\item{y_name}{The variable cible (Y) in characters.}

\item{x_name}{List of explanatory variables (list of characters).}

\item{alpha_}{0 for Ridge and 1 for Lasso.}

\item{lambda_cv}{Lambda min from \code{\link[=cv_glm_net]{cv_glm_net()}}}

\item{...}{Add arguments from \code{\link[=glmnet]{glmnet()}}.}
}
\description{
Fit a generalized linear model via penalized maximum likelihood \code{...}
}
\seealso{
\code{\link[=glmnet]{glmnet()}} for initial function

Other modeling functions: 
\code{\link{cv_glm_net}()},
\code{\link{def_model_keras}()},
\code{\link{fit_keras}()},
\code{\link{glm_logit}()},
\code{\link{tree_rpart}()}
}
\concept{modeling functions}
