% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam.R
\name{keyGen}
\alias{keyGen}
\title{create blank key structure}
\usage{
keyGen(numAges, numFleets)
}
\arguments{
\item{numAges}{number of ages}

\item{numFleets}{number of fleets}
}
\description{
Generate the blank key matrixces for filling
}
