% This is file ../spam0.29-3/man/methods.Rd
% This file is part of the spam package, 
%      http://www.math.uzh.ch/furrer/software/spam/
% written and maintained by Reinhard Furrer.
   










\name{spam methods}
\alias{diff.spam}
\title{Methods for sparse matrices}
\description{Methods without any additional parameters for sparse matrices.
}
\details{If a method for \code{spam} objects takes the same arguments,
  produces the intuitive output. We do not provide additional help
  pages.

  However, such methods are usually linked to a \code{xzy.spam}
  function, that could also be called directly.
	    }
\seealso{Corresponding base help functions.
}
\author{Reinhard Furrer}
\keyword{internal}
