% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam.R
\name{obsSetup}
\alias{obsSetup}
\title{Setup observation vector, aux stuff and idx}
\usage{
obsSetup(surveys, catch, landings, plusGroup)
}
\arguments{
\item{surveys}{list of survey fleets}

\item{catch}{list of catch fleets}

\item{landings}{landings statistics}

\item{plusGroup}{Form a plus group at some year?}
}
\description{
Setup observation vector, aux stuff and idx
}
