% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcasting_tools.R
\name{pred_glm_net}
\alias{pred_glm_net}
\title{Generalised predict function for \code{\link[=glmnet]{glmnet()}} models.}
\usage{
pred_glm_net(data_df, y_name, x_name, modele)
}
\arguments{
\item{data_df}{Data, class "data.frame".}

\item{y_name}{The variable cible (Y) in characters.}

\item{x_name}{List of explanatory variables (list of characters).}

\item{modele}{glmnet objects.}
}
\description{
Put data frame (train or test).
}
\seealso{
Other predicting functions: 
\code{\link{pred_rpart}()},
\code{\link{predict_broom}()}
}
\concept{predicting functions}
