% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcasting_tools.R
\name{pred_rpart}
\alias{pred_rpart}
\title{Vector of predictions for \code{\link[=rpart]{rpart()}} models.}
\usage{
pred_rpart(data_df, modele)
}
\arguments{
\item{data_df}{Data, class "data.frame".}

\item{modele}{rpart objects.}
}
\description{
Put data frame (train or test).
}
\seealso{
Other predicting functions: 
\code{\link{pred_glm_net}()},
\code{\link{predict_broom}()}
}
\concept{predicting functions}
