% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcasting_tools.R
\name{predict_broom}
\alias{predict_broom}
\title{Generalised predict function}
\usage{
predict_broom(modele, data, new_data, type = "response")
}
\arguments{
\item{modele}{broom-compatible model.}

\item{data}{model data, class "data.frame".}

\item{new_data}{Put data frame "test".}

\item{type}{Default at "response" or see \code{\link[=predict.glm]{predict.glm()}}, the type of prediction required.}
}
\description{
Generic function of prediction with function \code{\link[=augment]{augment()}} of package {broom}.
}
\examples{
library(broom)
fit <- glm(Ozone~., data= airquality)
predict_broom(modele= fit, data= airquality, new_data= airquality[2:50,], type= "response")
}
\seealso{
\code{\link[=augment]{augment()}} for initial function.

Other predicting functions: 
\code{\link{pred_glm_net}()}
}
\concept{predicting functions}
