% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{read_spam}
\alias{read_spam}
\title{Import des donnees "spam" (.txt).}
\usage{
read_spam(file)
}
\arguments{
\item{file}{Character containing path and name of file "spam".}
}
\value{
Data frame object.
}
\description{
La fonction est just prevue pour "spam".

Data input is path of file.
}
