% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tools.R
\name{rename_static}
\alias{rename_static}
\title{Reommage des variables (statique).}
\usage{
rename_static(data_frame)
}
\arguments{
\item{data_frame}{Data frame des donnees "SPAM".}
}
\value{
Le data frame mis a jour.
}
\description{
Les noms de colonnes deviennent "wf_" et "cf_".

Data input is data frame.
}
