% HEADER %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% This is file spam/man/spam.internal.Rd.                   %
% It is part of the R package spam,                         %
%  --> https://CRAN.R-project.org/package=spam              %
%  --> https://CRAN.R-project.org/package=spam64            %
%  --> https://git.math.uzh.ch/reinhard.furrer/spam         %
% by Reinhard Furrer [aut, cre], Florian Gerber [aut],      %
%    Roman Flury [aut], Daniel Gerber [ctb],                %
%    Kaspar Moesinger [ctb]                                 %
% HEADER END %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
   
\name{spam internal}
\alias{printSize}
\title{
  Spam internal and auxiliary functions
}
\description{
  The functions listed below are auxiliary functions and are not exported by the NAMESPACE. 
  The user should not require to call these directly. 
}
\details{\code{printSize(x, size=8, digits=2L)} constructs a string for
  pretty printing the object size \code{x}. \cr

}
\keyword{internal}
