% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_tools.R
\name{tree_rpart}
\alias{tree_rpart}
\title{Recursive Partitioning and Regression Trees}
\usage{
tree_rpart(data_df, y_name, x_name, ...)
}
\arguments{
\item{data_df}{Model data, class "data.frame".}

\item{y_name}{The variable cible (Y) in characters.}

\item{x_name}{List of explanatory variables (list of characters).}

\item{...}{Add arguments from \code{\link[=rpart]{rpart()}}.}
}
\description{
Fit a rpart model then prune.rpart {rpart} model.
}
\seealso{
\code{\link[=rpart]{rpart()}} and \code{\link[=prune]{prune()}}for initial function

Other modeling functions: 
\code{\link{cv_glm_net}()},
\code{\link{def_model_keras}()},
\code{\link{fit_keras}()},
\code{\link{glm_logit}()},
\code{\link{glm_net}()}
}
\concept{modeling functions}
