\name{spam history}
\alias{spam.history}
\alias{history.spam}
\alias{history}
\title{Brief overview of the history}
\description{Brief overview of \code{spam}s  history}
\usage{
  spam.history()
}
\details{This list is maintained for releases 0.05 and higher.
  \item{}{0.13: Minor bug fixes. }
  \item{}{0.12: Many internal changes. Start maintaining a proper
    CHANGELOG file.}
  \item{}{0.11: Considerable help file improvements. Changes in
    \code{c/rbind}, \code{chol}, print/summary methods and other minor
    improvements, first push to CRAN.}
  \item{}{0.10: Minor bug fixes and help file improvements.}
  \item{}{0.09: Proper NAMESPACE handling.}
  \item{}{0.08: \code{fields} uses \code{spam} now as a required
    package.
    The required linking functions are implemented in fields.}
  \item{}{0.07: Change of slot names.}
  \item{}{0.06: Implements \code{spam} and the function linking
            \code{fields} version 3.6. Last stable version before the
	    "merging" with fields. 
              Update to R2.4 and documentation fill-in.}
  \item{}{0.05: Implements \code{spam} and the function linking
              \code{fields} version 3.2. It should represent a fair working
              version, with large gaps in the documentation.}
}
\references{See also a more detailed description at
  \url{http://www.mines.edu/~rfurrer/software/spam/history.shtml}.}
\seealso{\code{\link{spam.todo}}.}
\keyword{internal}
