\name{spam-deprecated}
\alias{spam-deprecated}
\alias{subset.rows.spam}
\title{Deprecated Objects in Package \pkg{spam}}
\description{
  The functions or variables listed here are provided for compatibility
  with older versions of \R only, and may be defunct as soon as of the
  next release.
}
\usage{
  subset.rows.spam(...)
}
\arguments{
\item{...}{some arguments}
}
\seealso{
  \code{\link{Deprecated}},
  \code{\link{Defunct}}
}
\keyword{misc}